@@ -171,15 +172,15 @@
         "stylelint-config-prettier": "^8.0.2",
         "stylelint-config-recommended": "^3.0.0",
         "tailwindcss": "^2.2.7",
-        "ts-jest": "^26.2.0",
-        "ts-node": "^9.0.0",
-        "typescript": "^3.9.7",
+        "ts-jest": "^29.0.5",
+        "ts-node": "^10.9.1",
+        "typescript": "^4.9.5",
         "webpack": "^4.44.1",
         "webpack-bundle-analyzer": "^4.4.0",
         "webpack-cli": "^3.3.12",
         "webpack-dev-server": "^3.11.0",
         "webpack-merge": "^5.1.2",
-        "webpack-obfuscator": "^0.18.0"
+        "webpack-obfuscator": "^2.6.0"
     },
     "browserslist": [
         ">0.2%",
