@@ -1,47 +1,47 @@
-import { ExtendedAPIPlugin, DefinePlugin } from 'webpack';
-import MiniCssExtractPlugin from 'mini-css-extract-plugin';
-import merge from 'webpack-merge';
-import OptimizeCssAssetsPlugin from 'optimize-css-assets-webpack-plugin';
-import path from 'path';
-import CopyWebpackPlugin from 'copy-webpack-plugin';
-import JavaScriptObfuscator from 'webpack-obfuscator';
-const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
+import { ExtendedAPIPlugin, DefinePlugin } from "webpack";
+import MiniCssExtractPlugin from "mini-css-extract-plugin";
+import merge from "webpack-merge";
+import OptimizeCssAssetsPlugin from "optimize-css-assets-webpack-plugin";
+import path from "path";
+import CopyWebpackPlugin from "copy-webpack-plugin";
+const WebpackObfuscator = require("webpack-obfuscator");
+const BundleAnalyzerPlugin = require("webpack-bundle-analyzer").BundleAnalyzerPlugin;
 
-const rootDir = path.resolve(__dirname, '..');
-const BUILD_DIR = path.resolve(rootDir, 'build');
+const rootDir = path.resolve(__dirname, "..");
+const BUILD_DIR = path.resolve(rootDir, "build");
 
-import commonConfig from './common';
+import commonConfig from "./common";
 
-const domain = process.env.BUILD_DOMAIN ? process.env.BUILD_DOMAIN.split(',') : [];
+const domain = process.env.BUILD_DOMAIN ? process.env.BUILD_DOMAIN.split(",") : [];
 
 const plugins = [
     new ExtendedAPIPlugin(),
-    new DefinePlugin({ 'process.env.BUILD_EXPIRE': JSON.stringify(process.env.BUILD_EXPIRE) }),
+    new DefinePlugin({ "process.env.BUILD_EXPIRE": JSON.stringify(process.env.BUILD_EXPIRE) }),
     new OptimizeCssAssetsPlugin({
         assetNameRegExp: /\.css$/g,
-        cssProcessor: require('cssnano'),
+        cssProcessor: require("cssnano"),
         cssProcessorPluginOptions: {
-            preset: ['default', { discardComments: { removeAll: true } }],
+            preset: ["default", { discardComments: { removeAll: true } }],
         },
         canPrint: false,
     }),
     new CopyWebpackPlugin({
-        patterns: [{ from: 'public' }],
+        patterns: [{ from: "public" }],
     }),
-    new JavaScriptObfuscator({ rotateUnicodeArray: true, domainLock: domain }),
+    new WebpackObfuscator({ rotateStringArray: true, domainLock: domain }),
 ];
 
-if (process.env.ANALYZE === '1') {
+if (process.env.ANALYZE === "1") {
     plugins.push(new BundleAnalyzerPlugin());
 }
 
 const config = merge(commonConfig, {
-    mode: 'production',
+    mode: "production",
     output: {
         path: BUILD_DIR,
-        filename: '[name].[hash].js',
-        globalObject: 'this',
-        publicPath: '/',
+        filename: "[name].[hash].js",
+        globalObject: "this",
+        publicPath: "/",
     },
     optimization: {
         usedExports: false,
@@ -52,19 +52,19 @@ const config = merge(commonConfig, {
         rules: [
             {
                 test: /\.(css|sass|scss|pcss)$/,
-                use: ['style-loader', MiniCssExtractPlugin.loader, 'css-loader', 'sass-loader', 'postcss-loader'],
+                use: ["style-loader", MiniCssExtractPlugin.loader, "css-loader", "sass-loader", "postcss-loader"],
             },
             {
                 test: /\.(tsx|ts)?$/,
                 use: [
                     {
-                        loader: 'thread-loader',
+                        loader: "thread-loader",
                         options: {
                             poolTimeout: 2000,
                         },
                     },
                     {
-                        loader: 'ts-loader',
+                        loader: "ts-loader",
                         options: {
                             transpileOnly: true,
                             happyPackMode: true,
@@ -75,12 +75,25 @@ const config = merge(commonConfig, {
             },
             {
                 test: /\.(woff|woff2|eot|ttf|otf)$/i,
-                use: [{
-                    loader: 'file-loader',
-                    options: {
-                        outputPath: 'fonts',
+                use: [
+                    {
+                        loader: "file-loader",
+                        options: {
+                            outputPath: "fonts",
+                        },
                     },
-                }]
+                ],
+            },
+            {
+                test: /\.(js)?$/,
+                exclude: /node_modules/,
+                enforce: "post",
+                use: {
+                    loader: WebpackObfuscator.loader,
+                    options: {
+                        rotateStringArray: true,
+                    },
+                },
             },
         ],
     },
